project/app/requirements.txt
requests
flask
readability-lxml
beautifulsoup4